requirements.txt
streamlit==1.31.1
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.2.0
markdown==3.5.2
beautifulsoup4==4.12.3
python-frontmatter==1.1.0
PyGithub==2.2.0 
exa-py